separate menu class -rubocop pass-
