+++ feature_test.rb
@@ -1,7 +1,12 @@
-require './lib/menu.rb'
+# require './lib/menu.rb'
+# y = Menu.new("Chicken" => 5)
+# p y.show 
+
+
+require './lib/takeaway.rb'
 
 x = Takeaway.new
 x.show_menu
-x.select_meal("burger")
-x.calculate_fee
-x.order_total
+# x.select_meal("burger")
+# x.calculate_fee
+# x.order_total
