+++ lib/menu.rb
@@ -1,34 +1,12 @@
-class Takeaway
-  attr_reader :menu, :order, :total
+class Menu
+  attr_reader :meals
 
-  def initialize
-    @order = []
-    @total = 0
-  end
-  
-  def show_menu
-    @menu = {
-      "curry" => 12, 
-      "burger" => 14, 
-      "pizza" => 10
-    }
+  def initialize(meals)
+    @meals = meals
   end
 
-  def select_meal(choice)
-    @order.push(choice)
+  def display_menu
+    display = @meals.map do |name, price| "#{name}, £#{price}" end
+    display.join(", ")
   end
-
-  def order_total
-    @total
-  end
-
-  def calculate_fee
-    @menu.each do |key, value| 
-      if key.include? @order
-      @total = value.sum
-      else 
-      end
-    end
-  end
-
 end
