+++ spec/menu_spec.rb
@@ -1,32 +1,16 @@
-require 'menu'
+require 'menu.rb'
 
-describe Takeaway do
-  describe '#show_menu' do
-    it 'return the menu' do
-      expect(subject.show_menu).to eq subject.menu
-    end
-  end
-  
-  describe '#select_meal' do
-    let(:choice) { double :choice }
+describe Menu do
+  subject(:menu) { described_class.new(meals) }
 
-    it { is_expected.to respond_to(:select_meal).with(1).argument }
+  let(:meals) do { chicken: 5 } end
 
-    it 'stores the choices in an Order array' do
-      expect(subject.select_meal(choice)).to eq subject.order
-    end
+  it 'store a list of meals and prices' do 
+    expect(menu.meals).to eq(meals)
   end
 
-  describe '#order_total' do
-    let(:total) { double :total}
-    
-    it 'shows the total order cost' do
-      expect(subject.order_total).to eq :total #expected double, got nil
-    end
-  end
-
-  describe '#calculate_fee' do
-    it 'calculates a fee' do
-    end
+  it 'shows the menu as a list' do
+    x = "chicken, £5"
+    expect(menu.display_menu).to eq(x)
   end
 end
